+++ lib/domain/record/components/notification_bar/notification_bar_state_notifier.dart
